Refactor pipelines into individual folders (#284)
* Migrate pipelines to independent folders
* Update/add tests for new repository layout
* Add internal loading mechanism from environment variables
* Update Smart NVR for new layout
* Make Transportation2 functional
