+++ tools/visual-pipeline-and-platform-evaluation-tool/Dockerfile.vippet
@@ -30,8 +30,8 @@ ADD requirements.txt /home/dlstreamer/vippet/requirements.txt
 
 RUN pip install -r requirements.txt
 
-ADD diagrams/ /home/dlstreamer/vippet/diagrams
-
 ADD app.py optimize.py pipeline.py device.py explore.py benchmark.py utils.py /home/dlstreamer/vippet/
+
+ADD pipelines/ /home/dlstreamer/vippet/pipelines
 
 CMD ["python", "app.py"]
