+++ tools/visual-pipeline-and-platform-evaluation-tool/compose.yml
@@ -32,6 +32,7 @@ x-vippet: &vippet
   environment:
     GST_DEBUG: "1"
     TZ: America/Los_Angeles
+    PIPELINE: smartnvr
   volumes:
     - collector-signals:/home/dlstreamer/vippet/.collector-signals
     - ./models:/home/dlstreamer/vippet/models
