+++ tools/visual-pipeline-and-platform-evaluation-tool/pipelines/smartnvr/diagram.png
