+++ tools/visual-pipeline-and-platform-evaluation-tool/pipelines/transportation2/config.yaml
@@ -1,0 +1,5 @@
+metadata:
+name: Transportation Proxy Pipeline 2 (D-C)
+definition: The Transportation Proxy Pipeline 2 (D-C) collects road data to understand road usage,
+  and collects additional data for tracking vehicles. Also have applications in airports,
+  marine ports and railways from safety and security perspective.
